Inform there is exlpaination in next section
Inform the readers that there is exlpaination in next section. When I was reading this section without looking into the next section, I google for the mechanisms behind these rules and it costs me meaningless effort.

This is probabily because we do not have a title for each English section and I am not sure whether people reading the Chinese version find it clearly.